\name{R2DOC-package}
\alias{R2DOC-package}
\alias{R2DOC}
\docType{package}
\title{
Base package for R2'documents'
}
\description{
Base package for R2'documents' packages\cr
The package contains 'format' functions (\code{\link{tableProperties}}
, \code{\link{textProperties}}, \code{\link{parProperties}}, \code{\link{cellProperties}}) 
and tool functions for reporting.
}

\details{
\tabular{ll}{
Package: \tab R2DOC\cr
Type: \tab Package\cr
Version: \tab 0.3\cr
Date: \tab 2013-05-03\cr
License: \tab GPL-3\cr
Depends: \tab rJava, methods\cr
}



}
\seealso{
\code{\linkS4class{parProperties} object}\cr
\code{\linkS4class{textProperties} object}\cr
\code{\linkS4class{cellProperties} object}\cr
\code{\linkS4class{tableProperties} object}
}
\author{
Author: David Gohel <\email{david.gohel@gmail.com}; \email{david.gohel@lysis-consultants.fr}>\cr
Lysis-consultants \url{http://lysis-consultants.net}\cr
Maintainer: David Gohel <\email{david.gohel@gmail.com}; \email{david.gohel@lysis-consultants.fr}>\cr
Lysis-consultants \url{http://lysis-consultants.net}
}

\references{
W3schools: CSS Reference\cr
\url{http://www.w3schools.com/cssref/default.asp}
}

\keyword{package}
\keyword{addTable}
\keyword{replaceText}

